Minimal HTML stub.
Just a static page with local copies of D3, Underscore.js, jQuery, and
xerblin (which is a simple command interpreter I wrote.)

For now I am aiming to have the demo be self-contained in a directory so
anyone can just check out the repo and open index.html in their browser
and have it work, even if they aren't connected to the Internet.

I am also leaving the CSS and layout and such very much up to others.
I'll add just enough CSS to the site.css file to make it not look too
1985, but I'm no designer.